Moving registration of external functions
